deprecate some ModelCriteria methods
